rename folder and migrate string based ui state
